SmallGPT/requirements.txt
# Core ML Libraries
torch>=2.0.0
numpy>=1.24.0

# Data Processing
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Development
jupyter>=1.0.0
black>=23.0.0
pytest>=7.3.0

# Optional: For experiment tracking
wandb>=0.15.0

# Text processing
regex>=2023.5.5
